transform: fix computation of linear transform
Trying to decompose a linear transforms as
  rotate(a) * rotate(-b) * (1, x; y, 1) * rotate(b)
does not work, one degree of freedom is missing: By QR decomposition
with Q=rotate(a) * rotate(-b) and R=(1, x; y, 1) * rotate(b) we'd need
to be able to make an arbitrary upper triangular R. But if we create the
zero only two degrees of freedom are left - not enough.

Due to SVD it's clear that
  rotate(a) * rotate(-b) * (x, 0; 0, y) * rotate(b) can represent all
linear transformations.